Refactored tests organization
